+++ esp_cam_sensor/CHANGELOG.md
@@ -6,6 +6,7 @@
 - Added more output size formats to the OV3660 driver and fixed the configuration method for JPEG compression factor.
 - Added more output size formats to the OV3660 driver and fixed the configuration method for JPEG compression factor.
 - Added SC121AT MIPI YUV sensor driver.
+- Added SC1346 MIPI camera sensor driver.
 
 - Adjust AWB configuration in sc2336_default_p4_eco5.json for better white balance stability